Revert "Merge pull request #3740 from boardsesh/perf/2565-bottom-chrome-fanout"
This reverts commit 217824a193ff7fb262a6c1e2271f8159d879aff0, reversing
changes made to ab582de63103d7e94133448e446e08677159b0b7.
