+++ packages/mobile/src/hooks/use-bottom-chrome-metrics.ts
@@ -1,0 +1,78 @@
+import { useMemo } from 'react';
+import { useSegments } from 'expo-router';
+import { useWindowDimensions } from 'react-native';
+import { useSafeAreaInsets } from 'react-native-safe-area-context';
+import { useTheme } from '../providers/theme-provider';
+import { isAccessorySurfaceRoute, isTabsChromeRoute } from '../lib/route-segments';
+import { useStickyAccessoryPresence } from './use-sticky-accessory-presence';
+import { useNativeAccessoryActive, useNativeTabBar } from './use-bottom-accessory';
+import { useDeviceLayout } from './use-device-layout';
+import { computeBottomChromeMetrics } from './bottom-chrome-metrics';
+import { resolveDetailPaneSurface } from '../theme/size-class';
+import { SIDEBAR_WIDTH } from '../theme/layout';
+
+/**
+ * Reserves and offsets for chrome that floats over scrollable content (the
+ * persistent queue toolbar / iOS 26 bottom accessory). Gathers the React inputs
+ * and delegates the arithmetic to the pure {@link computeBottomChromeMetrics}.
+ */
+export function useBottomChromeMetrics() {
+  const insets = useSafeAreaInsets();
+  const segments = useSegments();
+  // Only the *presence* of a current climb matters here — subscribe to the
+  // presence-only selector, which flips solely when a climb appears/disappears.
+  // This keeps bottom chrome from re-rendering on queue mutations OR on
+  // climb-to-climb navigation across every screen that floats it. Wall-aware so a
+  // board-presence ("on the wall") climb counts too. Use the same sticky wrapper
+  // as the accessory mount gate so the JS-vs-native arbitration tracks what the
+  // accessory host actually shows (incl. its brief presence-blip hold).
+  const hasCurrentClimb = useStickyAccessoryPresence();
+  const { variant } = useTheme();
+  // The TAB BAR is present on every tab route, including pushed sub-routes (session
+  // detail keeps it). The player route counts too (it's a modal over the live tabs)
+  // so the tab-bar metrics don't churn across its open/close — see isTabsChromeRoute.
+  const insideTabs = isTabsChromeRoute(segments);
+  // The ACCESSORY only shows on a top-level tab page (plus occluded under the player).
+  // Keep it separate from `insideTabs` so a pushed sub-route still reserves tab-bar
+  // height but no longer reserves accessory/toolbar space for a bar that's gone.
+  const onAccessorySurface = isAccessorySurfaceRoute(segments);
+  const nativeAccessoryActive = useNativeAccessoryActive();
+  const nativeAccessoryMounted = onAccessorySurface && nativeAccessoryActive;
+  const nativeTabBar = useNativeTabBar();
+  const usesNativeTabBar = insideTabs && nativeTabBar;
+  // Regular-width iPad replaces the bottom tab bar with the left sidebar. Only
+  // widths that also mount the selected-climb detail pane suppress the floating
+  // queue toolbar; tight regular windows keep that toolbar because the pane is
+  // intentionally absent there.
+  const { width: windowWidth } = useWindowDimensions();
+  const { widthClass } = useDeviceLayout();
+  const usesSidebar = insideTabs && widthClass === 'regular';
+  const detailPaneOwnsQueue =
+    usesSidebar && resolveDetailPaneSurface({ width: windowWidth, widthClass, sidebarWidth: SIDEBAR_WIDTH }) === 'pane';
+
+  return useMemo(
+    () =>
+      computeBottomChromeMetrics({
+        uiVariant: variant,
+        usesNativeTabBar,
+        insetsBottom: insets.bottom,
+        insideTabs,
+        onAccessorySurface,
+        hasCurrentClimb,
+        nativeAccessoryMounted,
+        usesSidebar,
+        detailPaneOwnsQueue,
+      }),
+    [
+      variant,
+      usesNativeTabBar,
+      insets.bottom,
+      insideTabs,
+      onAccessorySurface,
+      hasCurrentClimb,
+      nativeAccessoryMounted,
+      usesSidebar,
+      detailPaneOwnsQueue,
+    ],
+  );
+}
